+++ packages/cli/src/api/formats/po.ts
@@ -24,7 +24,7 @@ const serialize = (items: CatalogType, options) =>
       item.msgid = key
       item.msgstr = [message.translation]
       item.comments = message.comments || []
-      item.extractedComments = message.comment ? [message.comment] : []
+      item.extractedComments = message.extractedComments || []
       if (options.origins) {
         item.references = message.origin ? message.origin.map(joinOrigin) : []
       }
@@ -53,12 +53,8 @@ const isObsolete = R.either(R.path(["flags", "obsolete"]), R.prop("obsolete"))
 const deserialize: (item: Object) => Object = R.map(
   R.applySpec({
     translation: R.compose(R.head, R.defaultTo([]), getTranslations),
-    comment: R.compose(R.head, R.defaultTo([]), getExtractedComments),
-    comments: (item) =>
-      R.concat(
-        getTranslatorComments(item) as string,
-        R.tail(getExtractedComments(item))
-      ),
+    extractedComments: R.compose(R.defaultTo([]), getExtractedComments),
+    comments: R.compose(R.defaultTo([]), getTranslatorComments),
     obsolete: isObsolete,
     origin: R.compose(R.map(splitOrigin), R.defaultTo([]), getOrigins),
     flags: getFlags,
@@ -66,6 +62,9 @@ const deserialize: (item: Object) => Object = R.map(
 )
 
 type POItem = InstanceType<typeof PO.Item>
+type PoFormatter = {
+  parse: (raw: string) => Object
+}
 
 const validateItems = R.forEach<POItem>((item) => {
   if (R.length(getTranslations(item)) > 1) {
@@ -78,7 +77,7 @@ const validateItems = R.forEach<POItem>((item) => {
 
 const indexItems = R.indexBy(getMessageKey)
 
-const po: CatalogFormatter = {
+const po: CatalogFormatter & PoFormatter = {
   catalogExtension: ".po",
 
   write(filename, catalog, options) {
@@ -103,7 +102,6 @@ const po: CatalogFormatter = {
     return this.parse(raw)
   },
 
-  // @ts-ignore
   parse(raw) {
     const po = PO.parse(raw)
     validateItems(po.items)