+++ packages/macro/index.d.ts
@@ -7,15 +7,16 @@ export function t(
   ...placeholders: any[]
 ): string
 
+export type UnderscoreDigit<T = string> = { [digit: string]: T }
 export type ChoiceOptions<T = string> = {
-  [digit: number]: T
   offset?: number
   zero?: T
   one?: T
   few?: T
   many?: T
   other?: T
-}
+} & UnderscoreDigit<T>
+
 export function plural(arg: number | string, options: ChoiceOptions): string
 export function selectOrdinal(
   arg: number | string,
@@ -39,10 +40,26 @@ export type ChoiceProps = {
 } & TransProps &
   ChoiceOptions<ReactNode>
 
+/**
+ * The types should be changed after this PR is merged
+ * https://github.com/Microsoft/TypeScript/pull/26797
+ *
+ * then we should be able to specify that key of values is same type as value.
+ * We would be able to remove separate type Values = {...} definition
+ * eg.
+ * type SelectProps<Values> = {
+ *  value?: Values
+ *  [key: Values]: string
+ * }
+ *
+ */
+type Values = { [key: string]: string }
+
 export type SelectProps = {
-  value?: string
-  other?: ReactNode
-} & TransProps
+  value: string
+  other: ReactNode
+} & TransProps &
+  Values
 
 export const Trans: ComponentType<TransProps>
 export const Plural: ComponentType<ChoiceProps>