+++ packages/macro/src/macroJs.ts
@@ -4,10 +4,11 @@ import { NodePath } from "@babel/traverse"
 
 import ICUMessageFormat from "./icu"
 import { zip, makeCounter } from "./utils"
-import { COMMENT, ID, MESSAGE } from "./constants"
+import { COMMENT, ID, MESSAGE, EXTRACT_MARK } from "./constants"
 
 const keepSpaceRe = /(?:\\(?:\r\n|\r|\n))+\s+/g
 const keepNewLineRe = /(?:\r\n|\r|\n)+\s+/g
+const removeExtraScapedLiterals = /(?:\\(.))/
 
 function normalizeWhitespace(text) {
   return text.replace(keepSpaceRe, " ").replace(keepNewLineRe, "\n").trim()
@@ -84,7 +85,7 @@ export default class MacroJs {
     // preserve line number
     newNode.loc = path.node.loc
 
-    this.addExtractMark(path)
+    path.addComment("leading", EXTRACT_MARK)
     // @ts-ignore
     path.replaceWith(newNode)
   }
@@ -98,7 +99,10 @@ export default class MacroJs {
       return
     }
 
-    if (this.types.isCallExpression(path.node) && this.isIdentifier(path.node.callee, "t")) {
+    if (
+      this.types.isCallExpression(path.node) &&
+      this.isIdentifier(path.node.callee, "t")
+    ) {
       this.replaceTAsFunction(path)
       return
     }
@@ -143,7 +147,6 @@ export default class MacroJs {
     this._expressionIndex = makeCounter()
 
     const descriptor = this.processDescriptor(path.node.arguments[0])
-    this.addExtractMark(path)
     path.replaceWith(descriptor)
   }
 
@@ -160,10 +163,6 @@ export default class MacroJs {
       ),
       [descriptor]
     )
-
-    this.addExtractMark(path)
-
-    // @ts-ignore
     path.replaceWith(newNode)
   }
 
@@ -185,6 +184,7 @@ export default class MacroJs {
    *
    */
   processDescriptor = (descriptor) => {
+    this.types.addComment(descriptor, "leading", EXTRACT_MARK)
     const messageIndex = descriptor.properties.findIndex(
       (property) => property.key.name === MESSAGE
     )
@@ -194,7 +194,12 @@ export default class MacroJs {
 
     // if there's `message` property, replace macros with formatted message
     const node = descriptor.properties[messageIndex]
-    const tokens = this.tokenizeNode(node.value, true)
+
+    // Inside message descriptor the `t` macro in `message` prop is optional.
+    // Template strings are always processed as if they were wrapped by `t`.
+    const tokens = this.types.isTemplateLiteral(node.value)
+      ? this.tokenizeTemplateLiteral(node.value)
+      : this.tokenizeNode(node.value, true)
 
     let messageNode = node.value
     if (tokens != null) {
@@ -265,7 +270,7 @@ export default class MacroJs {
 
           return {
             type: "text",
-            value,
+            value: this.clearBackslashes(value),
           }
         }),
         expressions: R.map((exp: babelTypes.Expression) =>
@@ -275,7 +280,6 @@ export default class MacroJs {
         ),
       }),
       (exp) => zip(exp.quasis, exp.expressions),
-      // @ts-ignore
       R.flatten,
       R.filter(Boolean)
     )
@@ -353,12 +357,10 @@ export default class MacroJs {
   }
 
   /**
-   * addExtractMark - add comment which marks the string/object
-   * for extraction.
-   * @lingui/babel-extract-messages looks for this comment
+   * We clean '//\` ' to just '`'
    */
-  addExtractMark = (path) => {
-    path.addComment("leading", "i18n")
+  clearBackslashes(value: string) {
+    return value.replace(removeExtraScapedLiterals, "`")
   }
 
   /**