+++ scripts/jest/config.unit.js
@@ -17,8 +17,10 @@ module.exports = {
     ".*.js.snap$",
   ],
   coverageReporters: ["html", "lcov", "text"],
+  modulePathIgnorePatterns: ["/build"],
 
   reporters: ["default", "jest-junit"],
+  setupFiles: ['set-tz/utc'],
   setupFilesAfterEnv: [require.resolve("./env.js")],
   snapshotSerializers: [
     "jest-serializer-html",